+++ CHANGELOG.md
@@ -6,7 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+
+### Fixed
+
+- Added support for running scripts with [yarn](https://yarnpkg.com/),
+  [pnpm](https://pnpm.io/), and older versions of npm.
 
 ## [0.2.0] - 2022-04-26